refactor(egress): drop duplicate fw-agent card; move env toggles to header top-right
The /egress page duplicated the Egress Firewall Agent host-singleton status
card already shown on /settings-egress-fw-agent, behind a collapsible
"Firewall agent & environments" section. Drops the card and collapsible,
and renders per-environment firewall toggles as a compact stacked list
in the page header's top-right (env name + switch). Confirm-disable
AlertDialog and the environment-egress-firewall-toggle data-tour anchor
are preserved; the now-unused EgressFirewallCard component is removed.
